requirements.txt
opencv-python
ultralytics
torch
flask
pyserial
numpy